About to implement lights migration

+++ Object_Oriented_Software_Development/Object_Oriented_Traffic_Problem/src/myproject/model/MP.java
@@ -5,11 +5,11 @@ package myproject.model;
  */
 public class MP {
 	private MP() {}
-	/** Length of cars, in meters */
-	public static double carLength = (int) (Math.random() * (ModelParameters.carLenMin - ModelParameters.carLenMax) + ModelParameters.carLenMin);
-	/** Length of roads, in meters */
-	public static double roadLength = (int) (Math.random() * (ModelParameters.roadLengthMin - ModelParameters.roadLengthMax) + ModelParameters.roadLengthMin);
-	/** Maximum car velocity, in meters/second */
-	public static double maxVelocity = (int) (Math.random() * (ModelParameters.carVelocityMin - ModelParameters.carVelocityMax) + ModelParameters.carVelocityMin);
+	/** Length of cars, in meters 10.0 */
+	public static double carLength = (Math.random() * (ModelParameters.carLenMin - ModelParameters.carLenMax) + ModelParameters.carLenMin);
+	/** Length of roads, in meters 200.0 */
+	public static double roadLength = (Math.random() * (ModelParameters.roadLengthMin - ModelParameters.roadLengthMax) + ModelParameters.roadLengthMin);
+	/** Maximum car velocity, in meters/second 6.0 */
+	public static double maxVelocity = (Math.random() * (ModelParameters.carVelocityMin - ModelParameters.carVelocityMax) + ModelParameters.carVelocityMin);
 }
